On branch api-refactor-doc  Changes to be committed: 	modified:   Procfile 	modified:   package.json 	modified:   public/index.html 	modified:   src/app.js 	modified:   src/controllers/authControllers.js 	modified:   src/controllers/rideControllers.js 	modified:   src/controllers/userControllers.js 	modified:   src/helpers/inputValidator.js 	modified:   src/helpers/seedData.js 	modified:   src/models/databaseMigrations.js 	new file:   src/models/productionDatabaseMigration.js 	modified:   src/routes/authRoutes.js 	modified:   src/routes/userRoutes.js 	modified:   src/test/auth.test.js 	modified:   src/test/rides.test.js 	modified:   src/test/users.test.js
FEATURE : PREVENT USERS FROM ACCEPTING OR REJECTING A REQUEST MADE TO
OTHERS RIDE AND LIMITING THE NUMBER OF REQUESTS A RIDE CAN TAKE

This prevent a user from accepting or rejecting requests not made to
their own ride offers. It also limit the number of requests a ride offer
can accept based on the number of sits available in the ride offer.

FINISHED #159133301